+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/pisdf/checker/PeriodsPreschedulingChecker.java
@@ -39,6 +39,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Set;
 import java.util.logging.Level;
 import org.eclipse.core.runtime.IProgressMonitor;
@@ -48,11 +49,15 @@ import org.preesm.commons.doc.annotations.PreesmTask;
 import org.preesm.commons.doc.annotations.Value;
 import org.preesm.commons.exceptions.PreesmRuntimeException;
 import org.preesm.commons.logger.PreesmLogger;
+import org.preesm.commons.model.PreesmCopyTracker;
 import org.preesm.model.pisdf.AbstractActor;
+import org.preesm.model.pisdf.AbstractVertex;
 import org.preesm.model.pisdf.Actor;
 import org.preesm.model.pisdf.ExecutableActor;
 import org.preesm.model.pisdf.PeriodicElement;
 import org.preesm.model.pisdf.PiGraph;
+import org.preesm.model.pisdf.brv.BRVMethod;
+import org.preesm.model.pisdf.brv.PiBRV;
 import org.preesm.model.scenario.PreesmScenario;
 import org.preesm.model.slam.Design;
 import org.preesm.workflow.elements.Workflow;
@@ -126,6 +131,23 @@ public class PeriodsPreschedulingChecker extends AbstractTaskImplementation {
           }
         }
       }
+    }
+
+    Map<AbstractVertex, Long> brv = PiBRV.compute(graph, BRVMethod.LCM);
+
+    for (final Entry<AbstractVertex, Long> en : brv.entrySet()) {
+      final AbstractVertex a = en.getKey();
+      AbstractVertex actor = PreesmCopyTracker.getOriginalSource(a);
+      long wcetMin = Long.MAX_VALUE;
+      for (final String operatorDefinitionID : scenario.getOperatorDefinitionIds()) {
+        final long timing = scenario.getTimingManager().getTimingOrDefault(actor.getVertexPath(), operatorDefinitionID)
+            .getTime();
+        if (timing < wcetMin) {
+          wcetMin = timing;
+        }
+      }
+      System.err.println(
+          "Actor : " + actor.getName() + " has wcet (min) : " + wcetMin + " [full path: " + actor.getVertexPath());
     }
 
     // 0. find all cycles and retrieve actors placed after delays.